简化代码，移动路由相关的功能到model模块，因为路由是由数据库决定的。
