+++ jpress/jpress-model/pom.xml
@@ -73,5 +73,12 @@
 		</dependency>
 
 
+		<dependency>
+			<groupId>javax.servlet</groupId>
+			<artifactId>javax.servlet-api</artifactId>
+			<scope>provided</scope>
+		</dependency>
+
+
 	</dependencies>
 </project>
