--- jpress/jpress-model/src/main/java/io/jpress/model/utils/PageRouter.java
@@ -1,32 +1,0 @@
-/**
- * Copyright (c) 2015-2016, Michael Yang 杨福海 (fuhai999@gmail.com).
- *
- * Licensed under the GNU Lesser General Public License (LGPL) ,Version 3.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.gnu.org/licenses/lgpl-3.0.txt
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.jpress.model.utils;
-
-import io.jpress.model.Content;
-
-public class PageRouter extends RouterConverter {
-
-
-	public static String getRouter(Content content) {
-		String url = SLASH + content.getSlug();
-
-		if (enalbleFakeStatic()) {
-			url += getFakeStaticSuffix();
-		}
-		return url;
-	}
-
-}
